\name{plotKML-package}
\alias{plotKML-package}
\alias{plotKML}
\docType{package}
\title{Plotting Spatial Data to KML}
\description{A suite of functions to facilitate authoring of KML documents from sp-class objects.}
\details{
\tabular{ll}{
Package: \tab plotKML\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-05-25\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
Tomislav Hengl <tom.hengl@wur.nl>, Dylan Beaudette <debeaudette@ucdavis.edu>
Maintainer: Tomislav Hengl <tom.hengl@wur.nl>
}
\references{}
\keyword{ package }
\seealso{}
\examples{}
